Make sysex.h compatible with C++
By using #ifdef __cplusplus and extern "C", it is ensured that the declarations in sysex.h are correctly handled by both C and C++ compilers. The explicit conversions with reinterpret_cast are necessary to inform the C++ compiler about the intended type conversion while ensuring type safety.

The C++ compatibility is needed for tools/regtool_osx.cpp (Port for MacOS), as it is based on the RtMidi.h/.cpp files as a replacement of alsa, which are used in regtool.c.