Rework OSC/MIDI dispatch code towards support of multiple devices
Each device will implement two functions, regtoctl and ctltoreg,
which convert betwen device register and generic control. Some
controls are parameterized by input and/or output channel, which
is set in struct param.

The OSC dispatch remains as a tree with the control ID set in the
leaf nodes, except we now call set() for each node towards the leaf.
At the "input" and "output" nodes, the set() function matches channel
numbers and sets param.in and param.out respectively. When a set()
handler needs to write a value, it calls setval(), which uses the
device's ctltoreg function to determine the register address.

The register dispatch now works differently. During initialization
the OSC tree is scanned, generating a mapping from control ID to
tree indices to get to that node. After parsing sysex data, registers
are converted to a control and parameters with regtoctl. Then, tree
indices are looked up for the control in the previously mentioned
table. Similar to OSC dispatch, we call new() for each node towards
the leaf. At the "input" and "output" nodes, we use the control
parameters to build the initial OSC address (e.g. "/input/3" or
"/output/7"). For other nodes, we just append the name to the node.
Finally, we have the full OSC address and call new() on the leaf
node.