Fix input stereo width calculation
When we get a new mix value, its given from the device as if the
left and right channels were routed individually to the output.
Accordingly, we adjust the mix as if the input channel were mono.
However, in this case calclevel() doesn't calculate l->width since
it only makes sense for stereo inputs.

The correct stereo mix width is determined later once we have applied
the updated value.